Started on basic melee attacker ai
