+++ Assets/Resources/Ai/StrikerBrain.controller
@@ -25,6 +25,18 @@ AnimatorStateTransition:
   m_InterruptionSource: 0
   m_OrderedInterruption: 1
   m_CanTransitionToSelf: 1
+--- !u!114 &-2481039319726712020
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e0c9d55359c6b344bafe1c85734e7b35, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1101 &-1777274075931708982
 AnimatorStateTransition:
   m_ObjectHideFlags: 1
@@ -91,7 +103,7 @@ AnimatorState:
   m_Transitions:
   - {fileID: 2690968225793427479}
   m_StateMachineBehaviours:
-  - {fileID: 4461151575438144997}
+  - {fileID: -2481039319726712020}
   m_Position: {x: 50, y: 50, z: 0}
   m_IKOnFeet: 0
   m_WriteDefaultValues: 1
@@ -120,10 +132,10 @@ AnimatorStateMachine:
     m_Position: {x: 50, y: 200, z: 0}
   - serializedVersion: 1
     m_State: {fileID: -1041954118178351125}
-    m_Position: {x: 370, y: 290, z: 0}
+    m_Position: {x: 350, y: 270, z: 0}
   - serializedVersion: 1
     m_State: {fileID: -1395173677252190428}
-    m_Position: {x: 120, y: 410, z: 0}
+    m_Position: {x: 160, y: 440, z: 0}
   m_ChildStateMachines: []
   m_AnyStateTransitions:
   - {fileID: 2520170947388132153}
@@ -237,18 +249,6 @@ AnimatorStateTransition:
   m_InterruptionSource: 0
   m_OrderedInterruption: 1
   m_CanTransitionToSelf: 1
---- !u!114 &4461151575438144997
-MonoBehaviour:
-  m_ObjectHideFlags: 1
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e6083e15bd4415e4eb5106879b9f17bb, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1102 &8887109603990110788
 AnimatorState:
   serializedVersion: 5
