+++ Assets/Scripts/Battlefield/StateBehaviors/PlayerAbilityStateBehavior.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fd5babead2fa83440aa55c62947ea5fe
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
